chore: clean up lingering interface definitions
