+++ src/services/cabins/__tests__/unit/interfaces.ts
@@ -1,5 +1,3 @@
-import { CabinBookingReceipt } from "@/services/mail/interfaces.js";
-
 import { BookingData } from "../../service.js";
 
 export interface NegativeValidationTestCase {
@@ -11,5 +9,8 @@ export interface NegativeValidationTestCase {
 export interface PositiveValidationTestCase {
   name: string;
   input: BookingData;
-  expectedConfirmationEmail: CabinBookingReceipt;
+  expectedConfirmationEmail: {
+    firstName: string;
+    lastName: string;
+  };
 }
