+++ src/services/cabins/service.ts
@@ -1,8 +1,9 @@
 import { Booking, Cabin } from "@prisma/client";
+import { MessageSendingResponse } from "postmark/dist/client/models/index.js";
 
 import { BookingStatus } from "@/domain/cabins.js";
 import { ValidationError } from "@/domain/errors.js";
-import { IMailService, TemplateAliasEnum } from "@/services/mail/interfaces.js";
+import { EmailContent, TemplateAlias } from "@/lib/postmark.js";
 
 import { bookingSchema } from "./validation.js";
 
@@ -29,6 +30,10 @@ export interface CabinRepository {
   }): Promise<Booking[]>;
 }
 
+export interface IMailService {
+  send(template: EmailContent): Promise<MessageSendingResponse>;
+}
+
 export class CabinService {
   constructor(
     private cabinRepository: CabinRepository,
@@ -45,7 +50,7 @@ export class CabinService {
 
   private sendBookingConfirmation(booking: Booking) {
     return this.mailService.send({
-      TemplateAlias: TemplateAliasEnum.CABIN_BOOKING_RECEIPT,
+      TemplateAlias: TemplateAlias.CABIN_BOOKING_RECEIPT,
       TemplateModel: {
         firstName: booking.firstName,
         lastName: booking.lastName,
