--- src/services/index.ts
@@ -1,3 +1,0 @@
-import { IAuthService, IMailService } from "./interfaces.js";
-
-export { IAuthService, IMailService };
